Moved the route for getting an opportunity's tags
