+++ backend/send_me/modules/tags/models.py
@@ -13,14 +13,6 @@ SQLAlchemy uses this to write the appropiate SQL
 for various operations.
 """
 
-
-# class Tags(Base):
-#     __table_name__ = "tags"
-
-#     id: Mapped[uuid.UUID] = mapped_column(primary_key=True)
-#     name: Mapped[str]
-
-
 class OpportunityTags(Base):
     __table_name__ = "opportunity_tags",
     opportunity_id:  Mapped[uuid.UUID] = mapped_column(ForeignKey("opportunities.id"), primary_key=True)
